EduGenie-main/requirements.txt
flask>=2.0
python-dotenv>=0.19
google-genai>=0.1.0
flask-cors>=3.0
